Revamped the whole structure of the site and apps
Changed from default template provided by Austin to one that allows for the displaying of the captcha and a simple text form. Still need to implement functionality of the button, and change the captcha displayer so that it displays from the folder (rn it just displays from the id i give it). need to find a way to point to the functions i wrote that moves and renames the captcha.