Added renaming functionality to site
Functions were already defined long ago, but now the submit button renames the file and sends it to the appropriate folder. also once all the captchas are renamed, it switches the view and gives a 'done' message. also instead of putting the image in a temp folder to avoid collision with multiple users at the same time, i made the selection random